+++ apps/storybook/.storybook/main.js
@@ -8,12 +8,10 @@ module.exports = {
     '../../../packages/menu/src/**/*.stories.js',
     '../../../packages/modal/src/**/*.stories.js',
     '../../../packages/popover/src/**/*.stories.js',
-    
-    '../../../packages/core/lib-esm/**/*.stories.mdx',
-    '../../../packages/core/lib-esm/**/*.stories.@(js|jsx|ts|tsx)',
-    
-    '../../../packages/slider/lib-esm/**/*.stories.mdx',
-    '../../../packages/slider/lib-esm/**/*.stories.@(js|jsx|ts|tsx)',
+
+    '../../../packages/core/src/**/*.stories.@(js|jsx|ts|tsx|mdx)',
+
+    '../../../packages/slider/src/**/*.stories.@(js|jsx|ts|tsx|mdx)',
   ],
   addons: storykitAddons,
   features,
