+++ packages/core/src/Button/Button.stories.mdx
@@ -1,0 +1,50 @@
+<!-- Checkbox.stories.mdx -->
+
+import { Canvas, Meta, Story } from '@storybook/addon-docs'
+import styled from 'styled-components'
+import { Button } from '.'
+import { ForwardRefs } from './ForwardRefsStory'
+import { argTypes } from './Button.stories.args'
+
+export const StyledButton = styled(Button)`
+  padding: 100px;
+`
+
+export const Template = (args) => <Button {...args}>I AM BUTTON</Button>
+export const StyledTemplate = (args) => <StyledButton {...args}>Button Text!</StyledButton>
+
+<Meta title='core/Button' component={Button} argTypes={argTypes} />
+
+# Button
+
+`Button` makes common actions immediately visible and easy to perform with one click or tap.
+
+<Canvas>
+  <Story name='Button'>{Template.bind({})}</Story>
+</Canvas>
+
+Buttons can have a responsive size based on the viewport width.
+
+<Canvas>
+  <Story
+    name='Responsive size'
+    args={{ size: ['small', null, 'medium', null, 'large'] }}
+    parameters={{
+      viewport: { defaultViewport: 'designSystem_sm' },
+    }}
+  >
+    {Template.bind({})}
+  </Story>
+</Canvas>
+
+If you need to access the `<button>` element directly, `Button` also has a `ref` prop that you can use to forward a ref to the underlying button.
+
+<Canvas>
+  <Story name='Forward refs'>{<ForwardRefs />}</Story>
+</Canvas>
+
+You can also use Styled Components to add custom styles to `Button`, e.g. `styled(Button)`.
+
+<Canvas>
+  <Story name='styled(Button) should preserve variation styling'>{StyledTemplate.bind({})}</Story>
+</Canvas>
