+++ tools/storybook-preset/src/index.ts
@@ -3,10 +3,10 @@ export * from '@storybook/addon-viewport'
 export * from 'storybook-addon-performance'
 
 /** @public */
-export const heftStoryGlobs = ['../lib-esm/**/*.stories.mdx', '../lib-esm/**/*.stories.@(js|jsx|ts|tsx)']
+export const heftStoryGlobs = ['../src/**/*.stories.@(js|jsx|ts|tsx|mdx)']
 
 /** @public */
-export const babelStoryGlobs = ['../src/**/*.stories.mdx', '../src/**/*.stories.@(js|jsx|ts|tsx)']
+export const babelStoryGlobs = ['../src/**/*.stories.@(js|jsx|ts|tsx|mdx)']
 
 /** @public */
 export const storykitAddons = [
